Improve project structure and documentation
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/imhuanglabau/tdp?shareId=XXXX-XXXX-XXXX-XXXX).
