+++ src/infrastructure/ui/components/MessageBox.jsx
