+++ src/infrastructure/ui/pages/MapPage.js
